Deprecate RPC getblocknumber
This RPC is exactly identical to getblockcount.  This duplication
dates back to commit 3c7ed41a0e511bf320979113b00c039f26098f93 when
Satoshi created the RPC interface.

There's no need to have both, so we standardize on "count" which
matches the naming convention in getconnectioncount.

Following the tradition established with previously deprecated APIs,
getblocknumber continues to work, but it's not listed in the help
system.