tests: Fix test_runner return value in case of skipped test
Summary:
Currently test_runner reports an error if a test case is skipped.
This is not how it should be, only failed tests should cause it to fail.

Backport of Core PR10187
https://github.com/bitcoin/bitcoin/pull/10187/files
This fixes the phantom failures that reported no actual error in CI.

Test Plan: Tested this patch on a CI machine

Reviewers: deadalnix, Fabien, #bitcoin_abc

Reviewed By: deadalnix, #bitcoin_abc

Subscribers: teamcity, schancel

Differential Revision: https://reviews.bitcoinabc.org/D2414