rpc: Add listen address to incoming connections in `getpeerinfo`
Summary:
 * rpc: Add listen address to incoming connections in `getpeerinfo`

This adds the listening address on which incoming connections were received to the
CNode and CNodeStats structures.

The address is reported in `getpeerinfo`.

This can be useful for distinguishing connections received on different listening ports
(e.g. when using a different listening port for Tor hidden service connections)
or different networks.

 * test: Add test for `getpeerinfo` `bindaddr` field

This is a backport of Core PR10478

Depends on D1750 and D1756

Test Plan:
  make check

Reviewers: #bitcoin_abc, schancel

Reviewed By: #bitcoin_abc, schancel

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1757