Decodehextx scripts sanity check
Summary:
 * Add a valid opcode sanity check to CScript

Added a function in CScript that checks if the script contains valid opcodes.

Add a test for that function

 * Sanity check transaction scripts in DecodeHexTx

Make sure that the scripts of decoded transactions are valid scripts.

This is a backport of Core PR10481

This include a bugfix to the integration test suite, as bogous transactions where generated in there.

Test Plan:
  make check
  ./test/functional/test_runner.py --extended

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D4247