Fix abc-p2p-activation.py when run on a monocore machine
Summary: Depending on the machine, parallel checking of script happens or not. When it does happen, the error code is different than when it doesn't. This update the test so that parallel checking is never used, and so the error code is always the same.

Test Plan: Run the test.

Reviewers: freetrader, kyuupichan, sickpig, awemany, #bitcoin_abc

Reviewed By: freetrader, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D203