Change default for excessive block size and generated block size
Summary:
Now that we ensure that the 1MB limit is respected before the fork is activated, we can safely increase defaults up to the value we want for the UAHF.

The generated block size is increased to 2MB and the maximum block size to 8MB by default.

Test Plan:
Updated tests to ensure the behavior is as expected.
Run the extended test suite to make sure they all pass.

Reviewers: freetrader, sickpig, awemany, kyuupichan, #bitcoin_abc

Reviewed By: freetrader, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D207