Increase the number of round we do before generating a new block template when CPU mining
Summary: When mining big block, the CreateNewBlock function takes longer to run. However, that function needs to lock cs_main . When running many thread on larger blocks, this cuases contention. Reducing the number of template we generate nicely avoid the contention.

Test Plan: Mining on test net. Without this patch, not all core are utilized at 100% because they are waiting on the lock. With the patch, it's all 100% .

Reviewers: freetrader, sickpig, awemany, kyuupichan, #bitcoin_abc

Reviewed By: freetrader, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D256