Create dependencies.md, and link it from README & build docs
Summary:
Backport of core PR10779 and PR11264
https://github.com/bitcoin/bitcoin/pull/10779/files
https://github.com/bitcoin/bitcoin/pull/11264/files

As OpenBSD is no longer supported the doc update is not applicable.

The versions used are set according to our depends system.
The minimum versions:
  - Clang and GCC: set to the first version with complete support for
C++14.
  - libevent: minimum 2.0.22 as 2.0.21 has a memory leak and would cause
asan to fail.
  - Qt: 5.5.1 is Ubuntu Xenial version, and matches core's master
minimum version. >= 5.4 is required to allow for `QTimer::singleShot` to
build.

Test Plan: Read it.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D3906