fixed travis build warning/error related to rename of package from bitcoin -> bitcoin-abc
Summary: Recent commit that fixed gitian builds to be 'bitcoin-abc' forgot to update travis.yml with new package name dir -- so the travis script was looking for a dir called build/bitcoin-$HOST when it should have been looking for build/bitcoin-abc-$HOST.  Updated

Test Plan:
Push this to github, enable travis builds, make sure building on travis works.  In case you want to see the result of my build testing this fix, it's here:

https://travis-ci.org/cculianu/bitcoin-abc/builds/268733782?utm_source=github_status&utm_medium=notification

A failed job (before this change) is here:

https://travis-ci.org/cculianu/bitcoin-abc/jobs/268729320#L5154

Notice error on line 5154.  Interestingly travis proceeds anyway but that red
error bothered me.

I locally use a travis-workalike set of scripts that I wrote so for me this
error to the .yml was fatal.

Reviewers: freetrader, deadalnix, #bitcoin_abc

Reviewed By: freetrader, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D489