Update OSX MIME type for .bitcoinpaymentrequest files, rename identifiers
Summary:
The OSX property list MIME type application/x-bitcoin-payment-request
is renamed to application/bitcoincash-paymentrequest in line with previous
changes in the wallet. This type is associated with .bitcoinpaymentrequest
files.

The identifiers used in the property list and payment protocol definition
file are updated to 'bitcoincash' to differentiate, and a few comments are
modified to reflect Bitcoin Cash.

No extensions of the existing payment protocol seem to have been
registered at the indicated Wiki URL, so the registration point has
been made into a TODO pending a decision on where such registrations
could occur.

This issue is marked as WIP pending review and confirmation of successful
tests.

Test Plan:
  make check (on wallet build)
  ../qa/pull-tester/rpc-tests.py -extended
Gitian builds
Travis builds

Reviewers: #bitcoin_abc, deadalnix, CCulianu

Reviewed By: #bitcoin_abc, CCulianu

Subscribers: CCulianu

Differential Revision: https://reviews.bitcoinabc.org/D508