Store the UTXO set on a per output basis rather than a per transaction basis
Summary:
This is the main chunk of the work started with D342 .

It come with an update on the database format and an upgrade routine to do the migration at startup if it is required.

Test Plan:
  make check
  ../qa/pull-tester/rpc-tests.py -extended

Also run a node on testnet from a folder that has been sync before this patch to test it upgrades properly.

Reviewers: freetrader, CCulianu, #bitcoin_abc, schancel

Reviewed By: #bitcoin_abc, schancel

Subscribers: hanchon, schancel, zquestz

Differential Revision: https://reviews.bitcoinabc.org/D512