wallet: Display non-HD error on first run
Summary:
```
On current master a fresh wallet created with -usehd=0 is silently
created as HD wallet.
An error should be displayed on the first run.

Also, this restores a test that was removed in c22a53c

Fixes: #11313
```

Backport of core PR11307
https://github.com/bitcoin/bitcoin/pull/11307/files

Test Plan:
  ./test/functional/test_runner.py wallet_hd

  mkdir /tmp/wallet
  ./src/bitcoind -walletdir=/tmp/wallet -usehd=0
Should return an init error:
```
Error: Error creating : You can't create non-HD wallets with this
version.
```

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D3150