[consensus] Pin P2SH activation to block 173805 on mainnet
Summary:
Partial backport of PR11389:
https://github.com/bitcoin/bitcoin/pull/11389/commits/18e071841e83044b47aa45c3e98c0796a407d445

Reviewers please also confirm that the rest of the PR is not worth
backporting (it has to do with segwit activation / versionbits), so we
can mark it 'done'.

This is not a totally faithful backport, due to divergence (D1262).

Depends on D3365 due to how the miner_tests build on a weird chain.

Test Plan:
`make check`
`test_runner.py --extended`
[x] IBD with -checkpoints=0 and -assumevalid=0 on mainnet: OK up to at least height 216000 (2013 January)
[x] IBD with -checkpoints=0 and -assumevalid=0 on testnet: OK up to at least height 230000 (2014 May)

Reviewers: jasonbcox, Fabien, #bitcoin_abc, deadalnix

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D3366