Add `-debuglogfile` option
Summary:
This patch adds an option to configure the name and/or directory of the
debug log.

The user can specify either a relative path, in which case the path
is relative to the data directory. They can also specify an absolute
path to put the log anywhere else in the file system.

Backport of core PR11781

Depends on D2393

Test Plan:
  ./test/functional/test_runner.py feature_logging

  bitcoind -debuglogfile=<path_to_debug_file>
Check the debug file is at the given location

  bitcoind -debuglogfile=<path_to_inexistant directory>
Check an error is output:
  Error: Could not open debug log file

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: schancel

Differential Revision: https://reviews.bitcoinabc.org/D2394