Add cmake/ninja build for the bitcoin-tx utility
Summary:
As per title.

I had to move some pieces around compared to the original design, but not a big deal because it wasn't making any sense anyway. It only works on the previous build system because of redundant linking.

Test Plan:
  ninja

Make sure the bitcoin-tx utility is built without errors.

Reviewers: schancel, #bitcoin_abc

Reviewed By: schancel, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D814