Implements a virtual destructor on the BaseRequestHandler class.
Summary:
```
Implements a virtual destructor on the BaseRequestHandler class to
protect against undefined behavior in the event that a derived
BaseRequestHandler class has a destructor and an object of such derived
class is destroyed through a pointer to its base class.
```

Backport of core PR12050
https://github.com/bitcoin/bitcoin/pull/12050/files

Note: the event described above actually occurs.

Test Plan:
Build with Clang and check there is no `-Wdelete-non-virtual-dtor`
warning.

With ASAN enabled:
  ./test/functional/test_runner.py interface_bitcoin_cli

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D3605