[qt] receive tab: bech32 address opt-in checkbox
Summary:
When launched with -adresstype=legacy the checkbox will be hidden.

This is a partial backport of Core PR11991 . The UI is not actually changed and getDefaultAddressType has been restored as PR10244 was improperly backported.

Test Plan:
  make check

Run bitcoin-qt and ensure the interface wasn't changed after clicking the recieve tab.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D3664