Activate replay protection at the time the next HF is scheduled.
Summary: This code will remain in place until the next HF is implemented, which ensure that there will be replay protection between the chains in case a minority survives by default. Obviously, the minority can chose to also do a hard fork and remove that replay protection, but then it is up to the minority chain promoter to convince actor in the space.

Test Plan: Added an integration test which grills the replay protection activation and deactivation (in case of reorg)

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox, teamcity

Maniphest Tasks: T301

Differential Revision: https://reviews.bitcoinabc.org/D1199