Add support to run the test suite via ninja
Summary:
This add the target check-bitcoin, check-univalue, check-leveldb, check-secp256k1 to run the respective test suites.

In addition, the magic target check is added to run all the test related to the project and the target check-all to run the test suite of our project and its dependencies.

Test Plan:
  ninja check
  ninja check-all
  ninja check-univalue
  ninja check-secp256k1
  ninja check-leveldb

Reviewers: #bitcoin_abc, schancel

Reviewed By: #bitcoin_abc, schancel

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1248