wallet: Reset BerkeleyDB handle after connection fails
Summary:
```
According to the BerkeleyDB docs, the DbEnv handle may not be accessed
after close() has been called. This change ensures that we create a new
handle after close() is called. This avoids a segfault when the first
connection attempt fails and then a second connection attempt tries to
call open() on the already closed DbEnv handle.
```

Backport of core PR13161.

Test Plan:
  ninja check
  ./test/functional/test_runner.py feature_dbcrash

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4883