[docs] Reformat -help output for help2man
Summary:
This commit slightly changes the format of the "Usage" strings in CLI
`-help` messages to meet the expection of the help2man tool, which we
use to generate man pages. On the way, we remove a few calls to
`strprintf()`, which became superficial after commit 32fbfda.

This is a backport of Core PR13872

Test Plan:
  make check
  BUILDDIR=~/bitcoin-abc/build ../contrib/devtools/gen-manpages.sh

Make sure the new generated manpages do make sense.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D3043